Corrects relationships between model-classes, make 1 file with all models
