+++ Makefile
@@ -1,5 +1,5 @@
 up:
-	docker compose -f docker-compose-local.yaml up -d
+	docker compose -f docker-compose-local.yaml up -d --remove-orphans
 
 down:
 	docker compose -f docker-compose-local.yaml down && docker network prune --force
